+++ tests/TomlBuilderTest.php
@@ -21,34 +21,34 @@ class TomlBuilderTest extends \PHPUnit_Framework_TestCase
     {
         $tb = new TomlBuilder();
         
-        $result = $tb->addComment('Toml file')->
-            addGroup('data.string')->
-            addValue('name', "Toml", 'This is your name')->
-            addValue('newline', "This string has a \n new line character.")->
-            addValue('winPath', "C:\\Users\\nodejs\\templates")->
-            addValue('unicode', 'unicode character: ' . json_decode('"\u03B4"'))->
+        $result = $tb->addComment('Toml file')
+            ->addTable('data.string')
+            ->addValue('name', "Toml", 'This is your name')
+            ->addValue('newline', "This string has a \n new line character.")
+            ->addValue('winPath', "C:\\Users\\nodejs\\templates")
+            ->addValue('unicode', 'unicode character: ' . json_decode('"\u03B4"'))
             
-            addGroup('data.bool')->
-            addValue('t', true)->
-            addValue('f', false)->
+            ->addTable('data.bool')
+            ->addValue('t', true)
+            ->addValue('f', false)
             
-            addGroup('data.integer')->
-            addValue('positive', 25, 'Comment inline.')->
-            addValue('negative', -25)->
+            ->addTable('data.integer')
+            ->addValue('positive', 25, 'Comment inline.')
+            ->addValue('negative', -25)
             
-            addGroup('data.float')->
-            addValue('positive', 25.25)->
-            addValue('negative', -25.25)->
+            ->addTable('data.float')
+            ->addValue('positive', 25.25)
+            ->addValue('negative', -25.25)
             
-            addGroup('data.datetime')->
-            addValue('datetime', new \Datetime())->
+            ->addTable('data.datetime')
+            ->addValue('datetime', new \Datetime())
             
-            addComment('Related to arrays')->
-            addGroup('data.array')->
-            addValue('simple', array(1,2,3))->
-            addValue('multiple', array( array(1,2), array('abc', 'def'), array(1.1, 1.2), array(true, false), array( new \Datetime()) ))->
-            
-            getTomlString();
+            ->addComment('Related to arrays')
+            ->addTable('data.array')
+            ->addValue('simple', array(1,2,3))
+            ->addValue('multiple', array( array(1,2), array('abc', 'def'), array(1.1, 1.2), array(true, false), array( new \Datetime()) ))
+
+            ->getTomlString();
 
         $this->assertNotNull(Toml::Parse($result));
     }
@@ -57,9 +57,9 @@ class TomlBuilderTest extends \PHPUnit_Framework_TestCase
     {
         $tb = new TomlBuilder();
         
-        $result = $tb->addComment('Toml file')->
-            addValue('thevoid', array())->
-            getTomlString();
+        $result = $tb->addComment('Toml file')
+            ->addValue('thevoid', array())
+            ->getTomlString();
 
         $this->assertNotNull(Toml::Parse($result));
     }
@@ -68,11 +68,11 @@ class TomlBuilderTest extends \PHPUnit_Framework_TestCase
     {
         $tb = new TomlBuilder();
         
-        $result = $tb->addGroup('a.b.c')->
-            addValue('answer', 42)->
-            addGroup('a')->
-            addValue('better', 43)->
-            getTomlString();
+        $result = $tb->addGroup('a.b.c')
+            ->addValue('answer', 42)
+            ->addTable('a')
+            ->addValue('better', 43)
+            ->getTomlString();
 
         $this->assertNotNull(Toml::Parse($result));
     }
@@ -81,11 +81,11 @@ class TomlBuilderTest extends \PHPUnit_Framework_TestCase
     {
         $tb = new TomlBuilder();
         
-        $result = $tb->addGroup('a')->
-            addValue('better', 43)->
-            addGroup('a.b.c')->
-            addValue('answer', 42)->
-            getTomlString();
+        $result = $tb->addGroup('a')
+            ->addValue('better', 43)
+            ->addTable('a.b.c')
+            ->addValue('answer', 42)
+            ->getTomlString();
 
         $this->assertNotNull(Toml::Parse($result));
     }
@@ -94,8 +94,8 @@ class TomlBuilderTest extends \PHPUnit_Framework_TestCase
     {
         $tb = new TomlBuilder();
         
-        $result = $tb->addGroup('a')->
-            getTomlString();
+        $result = $tb->addTable('a')
+            ->getTomlString();
 
         $this->assertNotNull(Toml::Parse($result));
     }
@@ -104,9 +104,9 @@ class TomlBuilderTest extends \PHPUnit_Framework_TestCase
     {
         $tb = new TomlBuilder();
         
-        $result = $tb->addGroup('a')->
-            addGroup('a.b')->
-            getTomlString();
+        $result = $tb->addGroup('a')
+            ->addTable('a.b')
+            ->getTomlString();
 
         $this->assertNotNull(Toml::Parse($result));
     }
@@ -115,8 +115,8 @@ class TomlBuilderTest extends \PHPUnit_Framework_TestCase
     {
         $tb = new TomlBuilder();
         
-        $result = $tb->addGroup('valid key')->
-            getTomlString();
+        $result = $tb->addGroup('valid key')
+            ->getTomlString();
 
         $this->assertNotNull(Toml::Parse($result));
     }
@@ -125,15 +125,15 @@ class TomlBuilderTest extends \PHPUnit_Framework_TestCase
     {
         $tb = new TomlBuilder();
         
-        $result = $tb->addValue('backspace', "This string has a \b backspace character.")->
-            addValue('tab', "This string has a \t tab character.")->
-            addValue('newline', "This string has a \n new line character.")->
-            addValue('formfeed', "This string has a \f form feed character.")->
-            addValue('carriage', "This string has a \r carriage return character.")->
-            addValue('quote', "This string has a \" quote character.")->
-            addValue('slash', "This string has a \/ slash character.")->
-            addValue('backslash', "This string has a \\ backslash character.")->
-            getTomlString();
+        $result = $tb->addValue('backspace', "This string has a \b backspace character.")
+            ->addValue('tab', "This string has a \t tab character.")
+            ->addValue('newline', "This string has a \n new line character.")
+            ->addValue('formfeed', "This string has a \f form feed character.")
+            ->addValue('carriage', "This string has a \r carriage return character.")
+            ->addValue('quote', "This string has a \" quote character.")
+            ->addValue('slash', "This string has a \/ slash character.")
+            ->addValue('backslash', "This string has a \\ backslash character.")
+            ->getTomlString();
 
         $this->assertNotNull(Toml::Parse($result));
     }
@@ -141,8 +141,8 @@ class TomlBuilderTest extends \PHPUnit_Framework_TestCase
     public function testKeySpecialChars() {
         $tb = new TomlBuilder();
         
-        $result = $tb->addValue("~!@#$^&*()_+-`1234567890[]\|/?><.,;:'", 1)->
-            getTomlString();
+        $result = $tb->addValue("~!@#$^&*()_+-`1234567890[]\|/?><.,;:'", 1)
+            ->getTomlString();
 
         $this->assertNotNull(Toml::Parse($result));
     }
@@ -150,17 +150,16 @@ class TomlBuilderTest extends \PHPUnit_Framework_TestCase
     public function testStringEscapesSingleQuote() {
         $tb = new TomlBuilder();
         
-        $result = $tb->addValue('backspace', 'This string has a \b backspace character.')->
-            addValue('tab', 'This string has a \t tab character.')->
-            addValue('newline', 'This string has a \n new line character.')->
-            addValue('formfeed', 'This string has a \f form feed character.')->
-            addValue('carriage', 'This string has a \r carriage return character.')->
-            addValue('quote', 'This string has a \" quote character.')->
-            addValue('slash', 'This string has a \/ slash character.')->
-            addValue('backslash', 'This string has a \\ backslash character.')->
-            getTomlString();
+        $result = $tb->addValue('backspace', 'This string has a \b backspace character.')
+            ->addValue('tab', 'This string has a \t tab character.')
+            ->addValue('newline', 'This string has a \n new line character.')
+            ->addValue('formfeed', 'This string has a \f form feed character.')
+            ->addValue('carriage', 'This string has a \r carriage return character.')
+            ->addValue('quote', 'This string has a \" quote character.')
+            ->addValue('slash', 'This string has a \/ slash character.')
+            ->addValue('backslash', 'This string has a \\ backslash character.')
+            ->getTomlString();
 
         $this->assertNotNull(Toml::Parse($result));
     }
-    
 }